[skip hazardlib][demos]
