+++ openquake/engine/engine.py
@@ -58,11 +58,6 @@ GET_JOBS = '''--- executing or submitted
 SELECT * FROM job WHERE status IN ('executing', 'submitted')
 AND is_running=1 AND pid > 0 ORDER BY id'''
 
-
-class TimeoutError(RuntimeError):
-    pass
-
-
 if OQ_DISTRIBUTE == 'zmq':
 
     def set_concurrent_tasks_default(job_id):
@@ -306,17 +301,6 @@ def run_calc(job_id, oqparam, exports, hazard_calculation_id=None, **kw):
     :param exports:
         A comma-separated string of export types.
     """
-    if OQ_DISTRIBUTE == 'zmq':  # start zworkers
-        master = w.WorkerMaster(config.dbserver.listen, **config.zworkers)
-        logs.dbcmd('start_zworkers', master)
-        for _ in range(60):
-            time.sleep(.5)
-            status = master.status()
-            if all(st == 'running' for host, st in status):
-                break
-        else:
-            raise TimeoutError(status)
-        logging.info('WorkerPool %s', status)
     register_signals()
     setproctitle('oq-job-%d' % job_id)
     calc = base.calculators(oqparam, calc_id=job_id)
@@ -349,6 +333,10 @@ def run_calc(job_id, oqparam, exports, hazard_calculation_id=None, **kw):
             del data  # save memory
 
         poll_queue(job_id, _PID, poll_time=15)
+        if OQ_DISTRIBUTE == 'zmq':  # start zworkers
+            master = w.WorkerMaster(config.dbserver.listen, **config.zworkers)
+            logs.dbcmd('start_zworkers', master)
+            logging.info('WorkerPool %s',  master.wait_pools())
         t0 = time.time()
         calc.run(exports=exports,
                  hazard_calculation_id=hazard_calculation_id,
